Move validation to separate module.
